finish task4

+++ TrainingAPI/app/apis/books_blueprint.py
@@ -14,8 +14,8 @@ from app.databases.mongodb import MongoDB
 from app.databases.redis_cached import get_cache, set_cache
 from app.decorators.json_validator import validate_with_jsonschema
 # from app.hooks.error import ApiInternalError
-from app.hooks.error import ApiInternalError
-from app.models.book import create_book_json_schema, Book
+from app.hooks.error import ApiInternalError, ApiNotFound
+from app.models.book import create_book_json_schema, Book, update_book_json_schema
 
 books_bp = Blueprint('books_blueprint', url_prefix='/books')
 
@@ -24,20 +24,15 @@ _db = MongoDB()
 
 @books_bp.route('/')
 async def get_all_books(request):
-    # # TODO: use cache to optimize api
-    # async with request.app.ctx.redis as r:
-    #     books = await get_cache(r, CacheConstants.all_books)
-    #     if books is None:
-    #         book_objs = _db.get_books()
-    #         books = [book.to_dict() for book in book_objs]
-    #         await set_cache(r, CacheConstants.all_books, books)
+    # TODO: use cache to optimize api
+    async with request.app.ctx.redis as r:
+        books = await get_cache(r, CacheConstants.all_books)
+        if books is None:
+            book_objs = _db.get_books()
+            books = [book.to_dict() for book in book_objs]
+            await set_cache(r, CacheConstants.all_books, books)
 
     book_objs = _db.get_books()
-
-    # query = {"_id": "a368b6ae-fb7c-48e7-a252-f1bd59decd1c"}
-    # book_obj = _db.get_book(query)
-    # find_book = [Book.to_dict(book_obj)]
-
     books = [book.to_dict() for book in book_objs]
     number_of_books = len(books)
     return json({
@@ -61,7 +56,14 @@ async def create_book(request, username=None):
     if not inserted:
         raise ApiInternalError('Fail to create book')
 
+    query = {"_id": "{}".format(book_id)}
+    inserted = _db.get_book(query)
+
     # TODO: Update cache
+    async with request.app.ctx.redis as r:
+        book_objs = _db.get_books()
+        books = [book.to_dict() for book in book_objs]
+        await set_cache(r, CacheConstants.all_books, books)
 
     return json({'status': 'success'})
 
@@ -71,32 +73,53 @@ async def create_book(request, username=None):
 async def get_book(request, _id: UUID):
     query = {"_id": "{}".format(_id)}
     book_obj = _db.get_book(query)
-    # print(book_obj)
+
+    if not book_obj:
+        raise ApiNotFound('id {}'.format(_id))
+
     return json({'book': book_obj})
 
 
-# @books_bp.route('/{id}', methods={'PUT'})
-# # @protected  # TODO: Authenticate
-# @validate_with_jsonschema(create_book_json_schema)  # To validate request body
-# async def update_book(request, username=None):
-#     body = request.json
-#
-#     book_id = str(uuid.uuid4())
-#     book = Book(book_id).from_dict(body)
-#     book.owner = username
-#
-#     # # TODO: Update book to database
-#     inserted = _db.update_book(book)
-#     if not inserted:
-#         raise ApiInternalError('Fail to update book')
-#
-#     # TODO: Update cache
-#
-#     return json({'status': 'success'})
+@books_bp.route('/<_id:uuid>', methods={'PUT'})
+# @protected  # TODO: Authenticate
+@validate_with_jsonschema(update_book_json_schema)  # To validate request body
+async def update_book(request, _id: UUID):
+    query = {"_id": "{}".format(_id)}
+    book_obj = _db.get_book(query)
+    if not book_obj:
+        raise ApiNotFound('id {}'.format(_id))
+
+    body = request.json
+
+    # TODO: Update book to database
+    book_obj = _db.update_book(query, body)
+    if not book_obj:
+        raise ApiInternalError('Fail to update book')
+    book_obj = _db.get_book(query)
+
+    # TODO: Update cache
+    async with request.app.ctx.redis as r:
+        book_objs = _db.get_books()
+        books = [book.to_dict() for book in book_objs]
+        await set_cache(r, CacheConstants.all_books, books)
+
+    return json({'book': book_obj})
 
 
 @books_bp.route('/<_id:uuid>', methods={'DELETE'})
 async def delete_book(request, _id: UUID):
     query = {"_id": "{}".format(_id)}
+    book_obj = _db.get_book(query)
+
+    if not book_obj:
+        raise ApiNotFound('id {}'.format(_id))
+
     book_obj = _db.delete_book(query)
+
+    # TODO: Update cache
+    async with request.app.ctx.redis as r:
+        book_objs = _db.get_books()
+        books = [book.to_dict() for book in book_objs]
+        await set_cache(r, CacheConstants.all_books, books)
+
     return json({'status': 'success'})
